Develop

+++ src/string/strrev.c
@@ -11,8 +11,9 @@
 
 char *strrev(char *str)
 {
-    for (size_t i = 0, j = strlen(str) - 1; i < j; ++i, --j) {
+    for (int i = 0, j = (int)(strlen(str)) - 1; i < j; ++i, --j) {
         char swap = str[i];
+
         str[i] = str[j];
         str[j] = swap;
     }
